Move content of new tests to old files
