+++ tests/integration/integration/relations/test_database.py
@@ -2,26 +2,23 @@
 # Copyright 2022 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-import asyncio
 import logging
-from pathlib import Path
 
+import jubilant_backports
 import pytest
-import yaml
-from pytest_operator.plugin import OpsTest
+from jubilant_backports import Juju
 
 from ... import markers
-from ...helpers import (
+from ...helpers_ha import (
+    CHARM_METADATA,
+    MINUTE_SECS,
     get_relation_data,
-    is_relation_broken,
-    is_relation_joined,
+    wait_for_apps_status,
 )
 
 logger = logging.getLogger(__name__)
 
-DB_METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
-
-DATABASE_APP_NAME = DB_METADATA["name"]
+DATABASE_APP_NAME = CHARM_METADATA["name"]
 DATABASE_ENDPOINT = "database"
 APPLICATION_APP_NAME = "mysql-test-app"
 APPLICATION_ENDPOINT = "database"
@@ -31,112 +28,94 @@ APPS = [DATABASE_APP_NAME, APPLICATION_APP_NAME]
 
 @pytest.mark.abort_on_fail
 @pytest.mark.skip_if_deployed
-async def test_build_and_deploy(ops_test: OpsTest, charm):
+def test_build_and_deploy(juju: Juju, charm):
     """Build the charm and deploy 3 units to ensure a cluster is formed."""
-    config = {"cluster-name": "test_cluster", "profile": "testing"}
-    resources = {"mysql-image": DB_METADATA["resources"]["mysql-image"]["upstream-source"]}
+    juju.deploy(
+        charm,
+        DATABASE_APP_NAME,
+        config={"cluster-name": "test_cluster", "profile": "testing"},
+        num_units=3,
+        resources={"mysql-image": CHARM_METADATA["resources"]["mysql-image"]["upstream-source"]},
+        base="ubuntu@22.04",
+        trust=True,
+    )
 
-    await asyncio.gather(
-        ops_test.model.deploy(
-            charm,
-            application_name=DATABASE_APP_NAME,
-            config=config,
-            num_units=3,
-            resources=resources,
-            base="ubuntu@22.04",
-            trust=True,
-        ),
-        ops_test.model.deploy(
-            APPLICATION_APP_NAME,
-            application_name=APPLICATION_APP_NAME,
-            num_units=2,
-            channel="latest/edge",
-            base="ubuntu@22.04",
-        ),
+    juju.deploy(
+        APPLICATION_APP_NAME,
+        num_units=2,
+        channel="latest/edge",
+        base="ubuntu@22.04",
     )
 
 
 @pytest.mark.abort_on_fail
-async def test_relation_creation_eager(ops_test: OpsTest):
+def test_relation_creation_eager(juju: Juju):
     """Relate charms before they have time to properly start.
 
     It simulates a Terraform-like deployment strategy
     """
-    await ops_test.model.relate(
+    logger.info("Creating relation...")
+    juju.integrate(
         f"{APPLICATION_APP_NAME}:{APPLICATION_ENDPOINT}",
         f"{DATABASE_APP_NAME}:{DATABASE_ENDPOINT}",
     )
-    await ops_test.model.block_until(
-        lambda: is_relation_joined(ops_test, APPLICATION_ENDPOINT, DATABASE_ENDPOINT) == True  # noqa: E712
+
+    logger.info("Waiting for application app to be waiting...")
+    juju.wait(
+        ready=wait_for_apps_status(jubilant_backports.all_waiting, APPLICATION_APP_NAME),
+        error=jubilant_backports.any_blocked,
+        timeout=15 * MINUTE_SECS,
     )
-
-    # Reduce the update_status frequency until the cluster is deployed
-    async with ops_test.fast_forward("60s"):
-        await ops_test.model.block_until(
-            lambda: len(ops_test.model.applications[DATABASE_APP_NAME].units) == 3
-        )
-        await ops_test.model.block_until(
-            lambda: len(ops_test.model.applications[APPLICATION_APP_NAME].units) == 2
-        )
-
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(
-                apps=[DATABASE_APP_NAME],
-                status="active",
-                raise_on_blocked=True,
-                timeout=1000,
-                raise_on_error=False,
-            ),
-            ops_test.model.wait_for_idle(
-                apps=[APPLICATION_APP_NAME],
-                status="waiting",
-                raise_on_blocked=True,
-                timeout=1000,
-            ),
-        )
+    logger.info("Waiting for database app to be active...")
+    juju.wait(
+        ready=wait_for_apps_status(jubilant_backports.all_active, DATABASE_APP_NAME),
+        error=jubilant_backports.any_blocked,
+        timeout=15 * MINUTE_SECS,
+    )
 
 
 @pytest.mark.abort_on_fail
 @markers.only_without_juju_secrets
-async def test_relation_creation_databag(ops_test: OpsTest):
+def test_relation_creation_databag(juju: Juju):
     """Relate charms and wait for the expected changes in status."""
-    async with ops_test.fast_forward("60s"):
-        await ops_test.model.wait_for_idle(apps=APPS, status="active")
+    juju.wait(
+        ready=jubilant_backports.all_active,
+        timeout=15 * MINUTE_SECS,
+    )
 
-    relation_data = await get_relation_data(ops_test, APPLICATION_APP_NAME, "database")
+    relation_data = get_relation_data(juju, APPLICATION_APP_NAME, "database")
     assert {"password", "username"} <= set(relation_data[0]["application-data"])
 
 
 @pytest.mark.abort_on_fail
 @markers.only_with_juju_secrets
-async def test_relation_creation(ops_test: OpsTest):
+def test_relation_creation(juju: Juju):
     """Relate charms and wait for the expected changes in status."""
-    async with ops_test.fast_forward("60s"):
-        await ops_test.model.wait_for_idle(apps=APPS, status="active")
+    juju.wait(
+        ready=jubilant_backports.all_active,
+        timeout=15 * MINUTE_SECS,
+    )
 
-    relation_data = await get_relation_data(ops_test, APPLICATION_APP_NAME, "database")
+    relation_data = get_relation_data(juju, APPLICATION_APP_NAME, "database")
     assert not {"password", "username"} <= set(relation_data[0]["application-data"])
     assert "secret-user" in relation_data[0]["application-data"]
 
 
 @pytest.mark.abort_on_fail
-async def test_relation_broken(ops_test: OpsTest):
+def test_relation_broken(juju: Juju):
     """Remove relation and wait for the expected changes in status."""
-    await ops_test.model.applications[DATABASE_APP_NAME].remove_relation(
+    juju.remove_relation(
         f"{APPLICATION_APP_NAME}:{APPLICATION_ENDPOINT}",
         f"{DATABASE_APP_NAME}:{DATABASE_ENDPOINT}",
     )
 
-    await ops_test.model.block_until(
-        lambda: is_relation_broken(ops_test, APPLICATION_ENDPOINT, DATABASE_ENDPOINT) == True  # noqa: E712
+    juju.wait(
+        ready=wait_for_apps_status(jubilant_backports.all_active, DATABASE_APP_NAME),
+        error=jubilant_backports.any_blocked,
+        timeout=15 * MINUTE_SECS,
     )
-
-    async with ops_test.fast_forward("60s"):
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(
-                apps=[DATABASE_APP_NAME], status="active", raise_on_blocked=True
-            ),
-            ops_test.model.wait_for_idle(
-                apps=[APPLICATION_APP_NAME], status="waiting", raise_on_blocked=True
-            ),
-        )
+    juju.wait(
+        ready=wait_for_apps_status(jubilant_backports.all_waiting, APPLICATION_APP_NAME),
+        error=jubilant_backports.any_blocked,
+        timeout=15 * MINUTE_SECS,
+    )
