+++ tests/integration/integration/roles/test_database_dba_role.py
@@ -2,96 +2,94 @@
 # Copyright 2025 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-import asyncio
 import logging
-from pathlib import Path
 
+import jubilant_backports
 import pytest
-import yaml
+from jubilant_backports import Juju
 from mysql.connector.errors import ProgrammingError
-from pytest_operator.plugin import OpsTest
 
-from ... import juju_
-from ...helpers import (
+from ...helpers_ha import (
+    CHARM_METADATA,
+    MINUTE_SECS,
     execute_queries_on_unit,
-    get_primary_unit,
+    get_app_units,
+    get_mysql_primary_unit,
     get_unit_address,
+    wait_for_apps_status,
 )
 
 logger = logging.getLogger(__name__)
 
-METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
-
-DATABASE_APP_NAME = METADATA["name"]
+DATABASE_APP_NAME = CHARM_METADATA["name"]
 INTEGRATOR_APP_NAME = "data-integrator"
 
 
 @pytest.mark.abort_on_fail
-async def test_build_and_deploy(ops_test: OpsTest, charm) -> None:
+def test_build_and_deploy(juju: Juju, charm) -> None:
     """Simple test to ensure that the mysql and data-integrator charms get deployed."""
-    resources = {"mysql-image": METADATA["resources"]["mysql-image"]["upstream-source"]}
-
-    async with ops_test.fast_forward("10s"):
-        await asyncio.gather(
-            ops_test.model.deploy(
-                charm,
-                application_name=DATABASE_APP_NAME,
-                num_units=3,
-                resources=resources,
-                base="ubuntu@22.04",
-                config={"profile": "testing"},
-            ),
-            ops_test.model.deploy(
-                INTEGRATOR_APP_NAME,
-                application_name=f"{INTEGRATOR_APP_NAME}1",
-                base="ubuntu@24.04",
-            ),
-            ops_test.model.deploy(
-                INTEGRATOR_APP_NAME,
-                application_name=f"{INTEGRATOR_APP_NAME}2",
-                base="ubuntu@24.04",
-            ),
-        )
-
-    await ops_test.model.wait_for_idle(
-        apps=[DATABASE_APP_NAME],
-        status="active",
+    juju.deploy(
+        charm,
+        DATABASE_APP_NAME,
+        num_units=3,
+        resources={"mysql-image": CHARM_METADATA["resources"]["mysql-image"]["upstream-source"]},
+        base="ubuntu@22.04",
+        config={"profile": "testing"},
     )
-    await ops_test.model.wait_for_idle(
-        apps=[f"{INTEGRATOR_APP_NAME}1", f"{INTEGRATOR_APP_NAME}2"],
-        status="blocked",
+    juju.deploy(
+        INTEGRATOR_APP_NAME,
+        f"{INTEGRATOR_APP_NAME}1",
+        base="ubuntu@24.04",
+    )
+    juju.deploy(
+        INTEGRATOR_APP_NAME,
+        f"{INTEGRATOR_APP_NAME}2",
+        base="ubuntu@24.04",
+    )
+
+    juju.wait(
+        ready=wait_for_apps_status(jubilant_backports.all_active, DATABASE_APP_NAME),
+        timeout=15 * MINUTE_SECS,
+    )
+    juju.wait(
+        ready=wait_for_apps_status(
+            jubilant_backports.all_blocked, f"{INTEGRATOR_APP_NAME}1", f"{INTEGRATOR_APP_NAME}2"
+        ),
+        timeout=15 * MINUTE_SECS,
     )
 
 
 @pytest.mark.abort_on_fail
-async def test_charmed_dba_role(ops_test: OpsTest):
+def test_charmed_dba_role(juju: Juju):
     """Test the database-level DBA role."""
-    await ops_test.model.applications[f"{INTEGRATOR_APP_NAME}1"].set_config({
-        "database-name": "preserved",
-        "extra-user-roles": "",
-    })
-    await ops_test.model.add_relation(f"{INTEGRATOR_APP_NAME}1", DATABASE_APP_NAME)
-    await ops_test.model.wait_for_idle(
-        apps=[f"{INTEGRATOR_APP_NAME}1", DATABASE_APP_NAME],
-        status="active",
+    juju.config(
+        f"{INTEGRATOR_APP_NAME}1",
+        {"database-name": "preserved", "extra-user-roles": ""},
+    )
+    juju.integrate(f"{INTEGRATOR_APP_NAME}1", DATABASE_APP_NAME)
+    juju.wait(
+        ready=wait_for_apps_status(
+            jubilant_backports.all_active, f"{INTEGRATOR_APP_NAME}1", DATABASE_APP_NAME
+        ),
+        timeout=15 * MINUTE_SECS,
     )
 
-    await ops_test.model.applications[f"{INTEGRATOR_APP_NAME}2"].set_config({
-        "database-name": "throwaway",
-        "extra-user-roles": "charmed_dba_preserved_00",
-    })
-    await ops_test.model.add_relation(f"{INTEGRATOR_APP_NAME}2", DATABASE_APP_NAME)
-    await ops_test.model.wait_for_idle(
-        apps=[f"{INTEGRATOR_APP_NAME}2", DATABASE_APP_NAME],
-        status="active",
+    juju.config(
+        f"{INTEGRATOR_APP_NAME}2",
+        {"database-name": "throwaway", "extra-user-roles": "charmed_dba_preserved_00"},
+    )
+    juju.integrate(f"{INTEGRATOR_APP_NAME}2", DATABASE_APP_NAME)
+    juju.wait(
+        ready=wait_for_apps_status(
+            jubilant_backports.all_active, f"{INTEGRATOR_APP_NAME}2", DATABASE_APP_NAME
+        ),
+        timeout=15 * MINUTE_SECS,
     )
 
-    mysql_unit = ops_test.model.applications[DATABASE_APP_NAME].units[0]
-    primary_unit = await get_primary_unit(ops_test, mysql_unit, DATABASE_APP_NAME)
-    primary_unit_address = await get_unit_address(ops_test, primary_unit.name)
-
-    data_integrator_2_unit = ops_test.model.applications[f"{INTEGRATOR_APP_NAME}2"].units[0]
-    results = await juju_.run_action(data_integrator_2_unit, "get-credentials")
+    primary_unit_name = get_mysql_primary_unit(juju, DATABASE_APP_NAME)
+    primary_unit_address = get_unit_address(juju, DATABASE_APP_NAME, primary_unit_name)
+    data_integrator_2_unit_name = get_app_units(juju, f"{INTEGRATOR_APP_NAME}2")[0]
+    results = juju.run(data_integrator_2_unit_name, "get-credentials").results
 
     logger.info("Checking that the database-level DBA role cannot create new databases")
     with pytest.raises(ProgrammingError):
